Add PyStackReg image alignment operator
This adds a multi-array aware image alignment operator that uses
PyStackReg.

By default, the operator uses a reference of the currently
selected slice index, if a slice module is present.

This has the capability to save/load transformation matrices.
Translation transformation matrices can also be applied to images
of different resolution.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>