Prevent possible segmentation fault
I had a segmentation fault one time where `unitsModified()` was called.
The only possible cause I can see is if the `newChildDataSource` was
deleted before it was called.

One way to fix it is to add `newChildDataSource` as a context object,
so that the offending signal will be disconnected if `newChildDataSource`
is destroyed. That is what is done here.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>