Make updates necessary to pyxrf-utils pixi.toml
The pixi installation would get stuck forever if `scikit-beam` is
located in the `pypi-dependencies`. Moving it to the conda-forge
dependencies fixes the issue.

This also adds a couple of apparently required dependencies from
databroker.

Finally, this makes the only platform Linux, since that is our
only target for using this package.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>