Sort ptycho files by angle
The displayed table of ptycho files is sorted by SID, which is good.

Usually, if the files are sorted by SID, that means they are also
sorted by angle. However, that is not always the case.

Sort the ptycho files by angle when stacking, as they need to be
sorted by angle in the final image stack.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>