Skip "Dwell" column when creating log file
The Dwell value is sometimes missing from the `param_input` array.
I don't have any idea why. But it's the last value, and we don't
use it, so we can just skip over it entirely.

This gets creating the log file to work properly when the "Dwell"
is missing for some reason.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>