Only check filtered angles for duplicates
We were previously also checking angles that were removed by the
filtering and thus were not going to be used. It did not make
sense to warn the user that duplicate angles were being used from
the angles that were being skipped over.

Now, duplicate angles are only checked from the angles which will
be used.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>