Perform a deep copy so progress updates work
It seems that somewhere unknown, the progress data is getting
deleted, even though it is supposed to be just a shallow copy. This
is causing crashes.

If we make a deep copy, it avoids the problem. It's probably not
that expensive to do, either, since the data object is empty.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>