Auto add volume visualization for live updates
This helps the user see the live updates much more easily. You used
to have to create a volume yourself to actually see the updates. So
I think this will enable a lot more users to utilize these live
update features more successfully.

It also fixes some strange crash that would happen if the data was
auto-updated but no visualization module had ever been created
for that child data source. I spent two hours trying to determine
the cause of the crash, but was not successful.

However, this update is needed anyways, and it fixes the crash.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>